Move install ID key to MetaKeys
