+++ collect_app/src/main/java/org/odk/collect/android/preferences/MetaKeys.java
@@ -3,6 +3,8 @@ package org.odk.collect.android.preferences;
 @SuppressWarnings("PMD.MissingStaticMethodInNonInstantiatableClass")
 public class MetaKeys {
 
+    public static final String KEY_INSTALL_ID = "metadata_installid";
+
     private MetaKeys() {
 
     }
